Update header implementation for upcoming Discourse changes
Discourse will soon be dropping the 'widget' system for rendering its header. This commit switches the theme from the deprecated APIs to the their modern equivalents. As a nice side effect, this also fixes the layout issue of the 'visit the curriculum' button for anonymous users.

More information on the changes can be found at https://meta.discourse.org/t/296544