Move curriculum-nav to before-header-panel to avoid positioning hacks
Previously it was being rendered as a button on the right hand side of the header, and then `position:fixed` was used to pull it to the left. This made for some weird rendering quirks. This commit uses a plugin outlet to render it in the correct part of the HTML, which removes the need for the positioning workarounds.