FIX: Increase header icon rule specificity (#121)
A side effect of the recent header upgrade was to introduce the `.btn` class on the header buttons. That means they are now subject to the `color: inherit` rule on line 557 of this theme's css.

This commit increases the specificity of the header icon selector so that it 'wins' over the more general `color: inherit` rule.

This should fix the icon color problems under the 'day' theme on freecodecamp.